qna/requirements.txt
azure-cognitiveservices-language-luis==0.6.0